fix: harden error classification, deduplicate schemas, decompose validation
Address issue #3 items 4, 5, and 6:

- Harden createUserError() with cause-chain traversal, statusCode support,
  instanceof checks for TypeError/SyntaxError, and tightened string matching
  patterns to reduce false positives (e.g., "403" → "403 Forbidden")
- Replace 24 manually duplicated JSON Schema objects in tools.ts files with
  zodInputSchema() helper using Zod v4's built-in toJSONSchema(), ensuring
  input schemas stay in sync with Zod validation schemas
- Decompose 991-line validation.ts monolith into validation/ directory with
  focused modules: security.ts, parsers.ts, checks.ts, extract.ts
