+++ src/tools/meta/tools.ts
@@ -4,6 +4,12 @@
  */
 
 import type { ExtendedToolDefinition, OutputSchema } from "../../types/index.js";
+import { zodInputSchema } from "../../utils/schema.js";
+import {
+  ListToolCategoriesInputSchema,
+  ListCategoryToolsInputSchema,
+  SuggestToolInputSchema,
+} from "./schemas.js";
 import { listToolCategories, listCategoryTools, suggestTool, setMetaTools } from "./handlers.js";
 
 // ============================================================================
@@ -112,16 +118,7 @@ export const metaTools: ExtendedToolDefinition[] = [
     name: "midnight-list-tool-categories",
     description:
       "📋 DISCOVERY TOOL: List available tool categories for progressive exploration. Use this FIRST to understand what capabilities are available, then drill into specific categories with midnight-list-category-tools. Reduces cognitive load by organizing 28 tools into 7 logical groups.",
-    inputSchema: {
-      type: "object" as const,
-      properties: {
-        includeToolCounts: {
-          type: "boolean",
-          description: "Include number of tools per category (default: true)",
-        },
-      },
-      required: [],
-    },
+    inputSchema: zodInputSchema(ListToolCategoriesInputSchema),
     outputSchema: listCategoriesOutputSchema,
     annotations: {
       readOnlyHint: true,
@@ -135,30 +132,7 @@ export const metaTools: ExtendedToolDefinition[] = [
     name: "midnight-list-category-tools",
     description:
       "📋 DISCOVERY TOOL: List tools within a specific category. Use after midnight-list-tool-categories to see detailed tool information for a category of interest. Supports progressive disclosure pattern.",
-    inputSchema: {
-      type: "object" as const,
-      properties: {
-        category: {
-          type: "string",
-          enum: [
-            "search",
-            "analyze",
-            "repository",
-            "versioning",
-            "generation",
-            "health",
-            "meta",
-            "compound",
-          ],
-          description: "Category to list tools for",
-        },
-        includeSchemas: {
-          type: "boolean",
-          description: "Include input/output schemas (default: false)",
-        },
-      },
-      required: ["category"],
-    },
+    inputSchema: zodInputSchema(ListCategoryToolsInputSchema),
     outputSchema: listCategoryToolsOutputSchema,
     annotations: {
       readOnlyHint: true,
@@ -184,16 +158,7 @@ USAGE GUIDANCE:
 • Call once with your intent - no need to call repeatedly
 • More specific intents get better matches
 • Use the primaryRecommendation for the best match`,
-    inputSchema: {
-      type: "object" as const,
-      properties: {
-        intent: {
-          type: "string",
-          description: "What you want to accomplish (natural language description)",
-        },
-      },
-      required: ["intent"],
-    },
+    inputSchema: zodInputSchema(SuggestToolInputSchema),
     outputSchema: suggestToolOutputSchema,
     annotations: {
       readOnlyHint: true,
