+++ src/tools/repository/tools.ts
@@ -4,6 +4,21 @@
  */
 
 import type { ExtendedToolDefinition } from "../../types/index.js";
+import { zodInputSchema } from "../../utils/schema.js";
+import {
+  GetFileInputSchema,
+  ListExamplesInputSchema,
+  GetLatestUpdatesInputSchema,
+  GetVersionInfoInputSchema,
+  CheckBreakingChangesInputSchema,
+  GetMigrationGuideInputSchema,
+  GetFileAtVersionInputSchema,
+  CompareSyntaxInputSchema,
+  GetLatestSyntaxInputSchema,
+  UpgradeCheckInputSchema,
+  FullRepoContextInputSchema,
+  ExtractContractStructureInputSchema,
+} from "./schemas.js";
 import {
   getFile,
   listExamples,
@@ -31,32 +46,7 @@ USAGE GUIDANCE:
 • Use 'ref' parameter to get specific versions (branch, tag, or commit)
 • Use startLine/endLine to request specific sections of large files
 • Files >50KB are truncated (first 25KB + last 25KB preserved)`,
-    inputSchema: {
-      type: "object" as const,
-      properties: {
-        repo: {
-          type: "string",
-          description: "Repository name (e.g., 'compact', 'midnight-js', 'example-counter')",
-        },
-        path: {
-          type: "string",
-          description: "File path within repository",
-        },
-        ref: {
-          type: "string",
-          description: "Branch, tag, or commit SHA (default: main)",
-        },
-        startLine: {
-          type: "number",
-          description: "Start line number (1-based, inclusive). Use to request specific sections.",
-        },
-        endLine: {
-          type: "number",
-          description: "End line number (1-based, inclusive). Use with startLine for a range.",
-        },
-      },
-      required: ["repo", "path"],
-    },
+    inputSchema: zodInputSchema(GetFileInputSchema),
     annotations: {
       readOnlyHint: true,
       openWorldHint: true,
@@ -69,17 +59,7 @@ USAGE GUIDANCE:
     name: "midnight-list-examples",
     description:
       "List available Midnight example contracts and DApps with descriptions, complexity ratings, and key features.",
-    inputSchema: {
-      type: "object" as const,
-      properties: {
-        category: {
-          type: "string",
-          enum: ["counter", "bboard", "token", "voting", "all"],
-          description: "Filter by example type (default: all)",
-        },
-      },
-      required: [],
-    },
+    inputSchema: zodInputSchema(ListExamplesInputSchema),
     annotations: {
       readOnlyHint: true,
       idempotentHint: true,
@@ -92,21 +72,7 @@ USAGE GUIDANCE:
     name: "midnight-get-latest-updates",
     description:
       "Retrieve recent changes and commits across Midnight repositories. Useful for staying up-to-date with the latest developments.",
-    inputSchema: {
-      type: "object" as const,
-      properties: {
-        since: {
-          type: "string",
-          description: "ISO date to fetch updates from (default: last 7 days)",
-        },
-        repos: {
-          type: "array",
-          items: { type: "string" },
-          description: "Specific repos to check (default: all configured repos)",
-        },
-      },
-      required: [],
-    },
+    inputSchema: zodInputSchema(GetLatestUpdatesInputSchema),
     annotations: {
       readOnlyHint: true,
       openWorldHint: true,
@@ -119,16 +85,7 @@ USAGE GUIDANCE:
     name: "midnight-get-version-info",
     description:
       "Get the latest version, release notes, and recent breaking changes for a Midnight repository. Use this to ensure you're working with the latest implementation.",
-    inputSchema: {
-      type: "object" as const,
-      properties: {
-        repo: {
-          type: "string",
-          description: "Repository name (e.g., 'compact', 'midnight-js', 'sdk')",
-        },
-      },
-      required: ["repo"],
-    },
+    inputSchema: zodInputSchema(GetVersionInfoInputSchema),
     annotations: {
       readOnlyHint: true,
       openWorldHint: true,
@@ -141,20 +98,7 @@ USAGE GUIDANCE:
     name: "midnight-check-breaking-changes",
     description:
       "Check if there are breaking changes between your current version and the latest release. Essential before upgrading dependencies.",
-    inputSchema: {
-      type: "object" as const,
-      properties: {
-        repo: {
-          type: "string",
-          description: "Repository name (e.g., 'compact', 'midnight-js')",
-        },
-        currentVersion: {
-          type: "string",
-          description: "Version you're currently using (e.g., 'v1.0.0', '0.5.2')",
-        },
-      },
-      required: ["repo", "currentVersion"],
-    },
+    inputSchema: zodInputSchema(CheckBreakingChangesInputSchema),
     annotations: {
       readOnlyHint: true,
       openWorldHint: true,
@@ -167,24 +111,7 @@ USAGE GUIDANCE:
     name: "midnight-get-migration-guide",
     description:
       "Get a detailed migration guide for upgrading between versions, including all breaking changes, deprecations, and recommended steps.",
-    inputSchema: {
-      type: "object" as const,
-      properties: {
-        repo: {
-          type: "string",
-          description: "Repository name (e.g., 'compact', 'midnight-js')",
-        },
-        fromVersion: {
-          type: "string",
-          description: "Version you're migrating from",
-        },
-        toVersion: {
-          type: "string",
-          description: "Target version (default: latest stable)",
-        },
-      },
-      required: ["repo", "fromVersion"],
-    },
+    inputSchema: zodInputSchema(GetMigrationGuideInputSchema),
     annotations: {
       readOnlyHint: true,
       openWorldHint: true,
@@ -197,24 +124,7 @@ USAGE GUIDANCE:
     name: "midnight-get-file-at-version",
     description:
       "Get the exact content of a file at a specific version. CRITICAL: Use this to ensure code recommendations match the user's version. Always prefer this over get-file when version accuracy matters.",
-    inputSchema: {
-      type: "object" as const,
-      properties: {
-        repo: {
-          type: "string",
-          description: "Repository name (e.g., 'compact', 'midnight-js')",
-        },
-        path: {
-          type: "string",
-          description: "File path within repository",
-        },
-        version: {
-          type: "string",
-          description: "Version tag (e.g., 'v1.0.0') or branch (e.g., 'main')",
-        },
-      },
-      required: ["repo", "path", "version"],
-    },
+    inputSchema: zodInputSchema(GetFileAtVersionInputSchema),
     annotations: {
       readOnlyHint: true,
       idempotentHint: true,
@@ -228,28 +138,7 @@ USAGE GUIDANCE:
     name: "midnight-compare-syntax",
     description:
       "Compare a file between two versions to see what changed. Use this before recommending code patterns to ensure they work with the user's version.",
-    inputSchema: {
-      type: "object" as const,
-      properties: {
-        repo: {
-          type: "string",
-          description: "Repository name (e.g., 'compact')",
-        },
-        path: {
-          type: "string",
-          description: "File path to compare",
-        },
-        oldVersion: {
-          type: "string",
-          description: "Old version tag (e.g., 'v0.9.0')",
-        },
-        newVersion: {
-          type: "string",
-          description: "New version tag (default: latest stable)",
-        },
-      },
-      required: ["repo", "path", "oldVersion"],
-    },
+    inputSchema: zodInputSchema(CompareSyntaxInputSchema),
     annotations: {
       readOnlyHint: true,
       idempotentHint: true,
@@ -270,16 +159,7 @@ Get the authoritative Compact syntax reference. Prevents hallucination by provid
 - Map.lookup()/Set.member() ARE available in circuits
 
 ALWAYS check this reference before writing Compact contracts.`,
-    inputSchema: {
-      type: "object" as const,
-      properties: {
-        repo: {
-          type: "string",
-          description: "Repository name (default: 'compact')",
-        },
-      },
-      required: [],
-    },
+    inputSchema: zodInputSchema(GetLatestSyntaxInputSchema),
     annotations: {
       readOnlyHint: true,
       openWorldHint: true,
@@ -297,20 +177,7 @@ ALWAYS check this reference before writing Compact contracts.`,
     name: "midnight-upgrade-check",
     description:
       "🚀 COMPOUND TOOL: Complete upgrade analysis in ONE call. Combines version check + breaking changes + migration guide. Use this instead of calling midnight-get-version-info, midnight-check-breaking-changes, and midnight-get-migration-guide separately. Saves ~60% tokens.",
-    inputSchema: {
-      type: "object" as const,
-      properties: {
-        repo: {
-          type: "string",
-          description: "Repository name (default: 'compact')",
-        },
-        currentVersion: {
-          type: "string",
-          description: "Your current version (e.g., 'v0.14.0', '0.13.5')",
-        },
-      },
-      required: ["currentVersion"],
-    },
+    inputSchema: zodInputSchema(UpgradeCheckInputSchema),
     outputSchema: {
       type: "object" as const,
       properties: {
@@ -362,24 +229,7 @@ ALWAYS check this reference before writing Compact contracts.`,
     name: "midnight-get-repo-context",
     description:
       "🚀 COMPOUND TOOL: Get everything needed to start working with a repository in ONE call. Combines version info + syntax reference + relevant examples. Use this at the start of a coding session instead of multiple individual calls. Saves ~50% tokens.",
-    inputSchema: {
-      type: "object" as const,
-      properties: {
-        repo: {
-          type: "string",
-          description: "Repository name (e.g., 'compact', 'midnight-js')",
-        },
-        includeExamples: {
-          type: "boolean",
-          description: "Include example code snippets (default: true)",
-        },
-        includeSyntax: {
-          type: "boolean",
-          description: "Include syntax reference (default: true)",
-        },
-      },
-      required: ["repo"],
-    },
+    inputSchema: zodInputSchema(FullRepoContextInputSchema),
     outputSchema: {
       type: "object" as const,
       properties: {
@@ -416,21 +266,7 @@ ALWAYS check this reference before writing Compact contracts.`,
       "Counter.value access, missing disclose() calls, potential overflow. " +
       "Use BEFORE generating contracts to catch syntax errors. " +
       "Note: Static analysis only - catches common patterns but not semantic errors.",
-    inputSchema: {
-      type: "object" as const,
-      properties: {
-        code: {
-          type: "string",
-          description: "The Compact contract source code to analyze (provide this OR filePath)",
-        },
-        filePath: {
-          type: "string",
-          description:
-            "Path to a .compact file to analyze (alternative to providing code directly)",
-        },
-      },
-      required: [],
-    },
+    inputSchema: zodInputSchema(ExtractContractStructureInputSchema),
     outputSchema: {
       type: "object" as const,
       properties: {
