+++ src/tools/repository/validation/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Contract validation module.
+ * Re-exports the main extractContractStructure function.
+ */
+export { extractContractStructure } from "./extract.js";
