+++ src/utils/index.ts
@@ -66,5 +66,8 @@ export type { HostedSearchResult, HostedSearchResponse, HostedSearchFilter } fro
 // Output serialization (YAML default, JSON optional)
 export { serialize, setOutputFormat, isJsonOutput, getOutputMimeType } from "./serializer.js";
 
+// Schema conversion utilities
+export { zodInputSchema } from "./schema.js";
+
 // Shared version constant
 export { CURRENT_VERSION } from "./version.js";
